Move errors to plugin pkg
