+++ plugin/errors.go
@@ -1,4 +1,4 @@
-package containerd
+package plugin
 
 import "errors"
